[NO-ISSUE] Fixing flaky tests (disabling keycloak dev services and LiveReloadProcessorTest) (#4211)
* [NO-ISSUE] Fixing flaky tests:
- Disabling unnecessary Keycloak Dev Services from tests
- Increased timeout for LiveReloadProcessorTest

* - missing disabled

* - Separated life reload tests

* - missing header
