+++ quarkus/addons/jwt-parser/integration-tests/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.keycloak.devservices.enabled=false
