+++ quarkus/addons/kubernetes/integration-tests/src/test/resources/application.properties
@@ -1,1 +1,2 @@
 quarkus.kubernetes-client.devservices.enabled=false
+quarkus.keycloak.devservices.enabled=false
