+++ quarkus/extensions/kogito-quarkus-serverless-workflow-extension/kogito-quarkus-serverless-workflow-integration-test/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.keycloak.devservices.enabled=false
