+++ quarkus/extensions/kogito-quarkus-serverless-workflow-jdbc-token-persistence-extension/kogito-quarkus-serverless-workflow-jdbc-token-persistence-integration-test/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.keycloak.devservices.enabled=false
